Forenkle varianter som ikke er i bruk etter migrering
