+++ src/main/java/no/nav/foreldrepenger/stønadskonto/grensesnitt/Stønadsdager.java
@@ -1,27 +1,47 @@
 package no.nav.foreldrepenger.stønadskonto.grensesnitt;
 
 import java.time.LocalDate;
+import java.util.Optional;
 
-import no.nav.foreldrepenger.stønadskonto.regelmodell.grunnlag.Brukerrolle;
 import no.nav.foreldrepenger.stønadskonto.regelmodell.grunnlag.Dekningsgrad;
+import no.nav.foreldrepenger.stønadskonto.regelmodell.konfig.Konfigurasjon;
+import no.nav.foreldrepenger.stønadskonto.regelmodell.konfig.Parametertype;
 
-public interface Stønadsdager {
+/*
+ * Beholdes til evt senere behov
+ */
+public class Stønadsdager {
 
-    static Stønadsdager instance(LocalDate regelverksdato) {
-        return new StønadsdagerImpl(regelverksdato);
+    private final LocalDate regelverksdato;
+
+    private Stønadsdager(LocalDate regelverksdato) {
+        this.regelverksdato = regelverksdato;
     }
 
+    public static Stønadsdager instance(LocalDate regelverksdato) {
+        return new Stønadsdager(regelverksdato);
+    }
 
-    int ekstradagerPrematur(LocalDate fødselsdato, LocalDate termindato);
+    public boolean endretFellesperiodeDekningsgrad80(Integer dager, LocalDate familieHendelseDato) {
+        if (dager == null || dager == 0) {
+            return false;
+        }return Konfigurasjon.STANDARD.getParameter(Parametertype.FELLESPERIODE_DAGER, Dekningsgrad.DEKNINGSGRAD_80, getRegelverksdato(familieHendelseDato)) != dager;
+    }
 
-    int ekstradagerFlerbarn(LocalDate familieHendelseDato, int antallBarn, Dekningsgrad dekningsgrad);
+    public boolean endretForeldrepengerDekningsgrad80(Integer dager, LocalDate familieHendelseDato) {
+        if (dager == null || dager == 0) {
+            return false;
+        }
+        return Konfigurasjon.STANDARD.getParameter(Parametertype.FORELDREPENGER_BARE_FAR_RETT_DAGER, Dekningsgrad.DEKNINGSGRAD_80, getRegelverksdato(familieHendelseDato)) != dager &&
+            Konfigurasjon.STANDARD.getParameter(Parametertype.FORELDREPENGER_MOR_ELLER_ALENEOMSORG_DAGER, Dekningsgrad.DEKNINGSGRAD_80, getRegelverksdato(familieHendelseDato)) != dager;
+    }
 
-    int andredagerFarRundtFødsel(LocalDate hendelsedato, boolean gjelderFødsel);
+    private LocalDate getRegelverksdato(LocalDate familieHendelseDato) {
+        return Optional.ofNullable(regelverksdato).orElse(familieHendelseDato);
+    }
 
-    int minsterettTetteFødsler(Brukerrolle brukerrolle, boolean gjelderFødsel, LocalDate familiehendelse, LocalDate familiehendelseNesteSak);
-
-    int minsterettBareFarRett(LocalDate familiehendelse, int antallBarn, boolean bareFarRett, boolean morAvklartUfør, Dekningsgrad dekningsgrad);
-
-    int aktivitetskravUføredager(LocalDate familiehendelse, boolean bareFarRett, boolean morAvklartUfør, Dekningsgrad dekningsgrad);
+    private LocalDate getRegelverksdato(LocalDate familieHendelseDato1, LocalDate familieHendelseDato2) {
+        return Optional.ofNullable(regelverksdato).or(() -> Optional.ofNullable(familieHendelseDato1)).orElse(familieHendelseDato2);
+    }
 
 }
