+++ src/main/java/no/nav/foreldrepenger/stønadskonto/regelmodell/StønadskontoRegelOrkestrering.java
@@ -1,7 +1,5 @@
 package no.nav.foreldrepenger.stønadskonto.regelmodell;
 
-import java.util.Map;
-
 import no.nav.foreldrepenger.stønadskonto.regelmodell.grunnlag.BeregnKontoerGrunnlag;
 import no.nav.foreldrepenger.stønadskonto.regelmodell.regler.BeregnKontoer;
 import no.nav.foreldrepenger.stønadskonto.regelmodell.regler.KontoerMellomregning;
@@ -25,21 +23,8 @@ public class StønadskontoRegelOrkestrering {
 
         var førFletting = mellomregning.getBeregnet();
         var stønadskontoer = mellomregning.getFlettet().isEmpty() ? førFletting : mellomregning.getFlettet();
-        var stønadkontoerBeholdStønadsdager = mellomregning.getFlettetBeholdStønadsdager().isEmpty() ?
-            stønadskontoer : mellomregning.getFlettetBeholdStønadsdager();
-        var antallFlerbarnsdager = hentAntallFlerbarnsdager(stønadskontoer);
-        var antallPrematurDager = hentAntallPrematurDager(stønadskontoer);
 
-        return new StønadskontoResultat(stønadskontoer, stønadkontoerBeholdStønadsdager, førFletting,
-            antallFlerbarnsdager, evaluationJson, grunnlagJson, antallPrematurDager);
-    }
-
-    private int hentAntallFlerbarnsdager(Map<StønadskontoKontotype, Integer> kontoer) {
-        return kontoer.getOrDefault(StønadskontoKontotype.TILLEGG_FLERBARN, 0);
-    }
-
-    private int hentAntallPrematurDager(Map<StønadskontoKontotype, Integer> kontoer) {
-        return kontoer.getOrDefault(StønadskontoKontotype.TILLEGG_PREMATUR, 0);
+        return new StønadskontoResultat(stønadskontoer, evaluationJson, grunnlagJson);
     }
 
     private String toJson(BeregnKontoerGrunnlag grunnlag) {
