+++ src/main/java/no/nav/foreldrepenger/stønadskonto/regelmodell/regler/FletteKontoer.java
@@ -27,15 +27,6 @@ class FletteKontoer extends LeafSpecification<KontoerMellomregning> {
         var tidligereBeregnet = new EnumMap<>(mellomregning.getGrunnlag().getTidligereUtregning());
 
         /*
-         * For utprøving ved migrering i fpsak. Mulig denne blir default
-         * Beholder input-verdi av stønadsdager, utvidelser og flerbarnsdager. Fersk utregning av øvrige
-         */
-        Map<StønadskontoKontotype, Integer> flettetBeholdStønad = new EnumMap<>(StønadskontoKontotype.class);
-        Arrays.stream(StønadskontoKontotype.values())
-            .forEach(konto -> flettVerdi(konto, FletteKontoer::lhs, tidligereBeregnet.get(konto), beregnet.get(konto)).ifPresent(v -> flettetBeholdStønad.put(konto, v)));
-        mellomregning.getFlettetBeholdStønadsdager().putAll(flettetBeholdStønad);
-
-        /*
          * Stønadsdager, utvidelser og flerbarnsdager: Beholder opprinnelig dersom ulik konfig, ellers beregning med maxverdi
          * Øvrige retter: Bruk alltid siste utregning.
          *
@@ -47,15 +38,6 @@ class FletteKontoer extends LeafSpecification<KontoerMellomregning> {
             .forEach(konto -> flettMax(konto, kontovelgerMax, tidligereBeregnet.get(konto), beregnet.get(konto)).ifPresent(v -> flettetBrukMaxStønad.put(konto, v)));
         mellomregning.getFlettet().putAll(flettetBrukMaxStønad);
         return ja();
-    }
-
-    private static Optional<Integer> flettVerdi(StønadskontoKontotype konto, BiFunction<Integer, Integer, Optional<Integer>> kontovelger,
-                                               Integer opprinnelig, Integer beregnet) {
-        return switch (konto) {
-            case FELLESPERIODE, MØDREKVOTE, FEDREKVOTE, FORELDREPENGER, FORELDREPENGER_FØR_FØDSEL,
-                TILLEGG_FLERBARN, TILLEGG_PREMATUR, FLERBARNSDAGER -> kontovelger.apply(opprinnelig, beregnet).filter(v -> v > 0);
-            case UFØREDAGER, TETTE_SAKER_MOR, TETTE_SAKER_FAR, BARE_FAR_RETT, FAR_RUNDT_FØDSEL -> Optional.ofNullable(beregnet).filter(v -> v > 0);
-        };
     }
 
     private static Optional<Integer> flettMax(StønadskontoKontotype konto, BiFunction<Integer, Integer, Optional<Integer>> kontovelger,
