+++ src/test/java/no/nav/foreldrepenger/stønadskonto/regelmodell/TidligereUtregningTest.java
@@ -40,9 +40,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(FORELDREPENGER, 230))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FORELDREPENGER, 230);
-        assertThat(retter).doesNotContainKey(FELLESPERIODE);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FORELDREPENGER, 230);
         assertThat(retterMax).doesNotContainKey(FELLESPERIODE);
@@ -58,9 +55,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(MØDREKVOTE, 75, FEDREKVOTE, 75, FELLESPERIODE, 80, FORELDREPENGER_FØR_FØDSEL, 15))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FELLESPERIODE, 80);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FELLESPERIODE, 80);
         assertThat(retterMax).doesNotContainKey(FORELDREPENGER);
@@ -77,10 +71,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(MØDREKVOTE, 75, FEDREKVOTE, 75, FELLESPERIODE, 80, FORELDREPENGER_FØR_FØDSEL, 15))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FELLESPERIODE, 80);
-        assertThat(retter).doesNotContainKey(StønadskontoKontotype.TILLEGG_PREMATUR);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FELLESPERIODE, 120);
         assertThat(retterMax).containsEntry(StønadskontoKontotype.TILLEGG_PREMATUR, 40);
@@ -98,10 +88,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(MØDREKVOTE, 75, FEDREKVOTE, 75, FELLESPERIODE, 120, FORELDREPENGER_FØR_FØDSEL, 15))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FELLESPERIODE, 120);
-        assertThat(retter).containsEntry(FAR_RUNDT_FØDSEL, 10);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FELLESPERIODE, 120);
         assertThat(retterMax).containsEntry(FAR_RUNDT_FØDSEL, 10);
@@ -118,10 +104,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(MØDREKVOTE, 75, FEDREKVOTE, 75, FELLESPERIODE, 80, FORELDREPENGER_FØR_FØDSEL, 15))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FELLESPERIODE, 80);
-        assertThat(retter).containsEntry(MØDREKVOTE, 75);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FELLESPERIODE, 90);
         assertThat(retterMax).containsEntry(MØDREKVOTE, 95);
@@ -139,11 +121,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(FORELDREPENGER, 200))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FORELDREPENGER, 200);
-        assertThat(retter).containsEntry(UFØREDAGER, 75);
-        assertThat(retter).doesNotContainKey(FELLESPERIODE);
-        assertThat(retter).doesNotContainKey(BARE_FAR_RETT);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FORELDREPENGER, 200);
         assertThat(retterMax).containsEntry(UFØREDAGER, 75);
@@ -163,11 +140,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(FORELDREPENGER, 200))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FORELDREPENGER, 200);
-        assertThat(retter).containsEntry(BARE_FAR_RETT, 75);
-        assertThat(retter).doesNotContainKey(FELLESPERIODE);
-        assertThat(retter).doesNotContainKey(UFØREDAGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FORELDREPENGER, 200);
         assertThat(retterMax).containsEntry(BARE_FAR_RETT, 75);
@@ -186,12 +158,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(MØDREKVOTE, 75, FEDREKVOTE, 75, FELLESPERIODE, 80, FORELDREPENGER_FØR_FØDSEL, 15))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FELLESPERIODE, 80);
-        assertThat(retter).containsEntry(MØDREKVOTE, 75);
-        assertThat(retter).containsEntry(TETTE_SAKER_MOR, 110);
-        assertThat(retter).containsEntry(TETTE_SAKER_FAR, 40);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FELLESPERIODE, 90);
         assertThat(retterMax).containsEntry(MØDREKVOTE, 95);
@@ -211,12 +177,6 @@ class TidligereUtregningTest {
             .tidligereUtregning(Map.of(FORELDREPENGER, 280))
             .build();
         var stønadskontoResultat = stønadskontoRegelOrkestrering.beregnKontoer(grunnlag);
-        var retter = stønadskontoResultat.getStønadskontoerBeholdStønadsdager();
-        assertThat(retter).containsEntry(FORELDREPENGER, 280);
-        assertThat(retter).containsEntry(TETTE_SAKER_MOR, 40);
-        assertThat(retter).doesNotContainKey(TETTE_SAKER_FAR);
-        assertThat(retter).doesNotContainKey(FELLESPERIODE);
-        assertThat(retter).doesNotContainKey(FORELDREPENGER_FØR_FØDSEL);
         var retterMax = stønadskontoResultat.getStønadskontoer();
         assertThat(retterMax).containsEntry(FORELDREPENGER, 280);
         assertThat(retterMax).containsEntry(TETTE_SAKER_MOR, 40);
